requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
psycopg2==2.9.10
SQLAlchemy==2.0.39
Werkzeug==3.1.3
Jinja2==3.1.6
itsdangerous==2.2.0
openai==1.68.2
python-dotenv==0.20.0
pdf2image
pdfplumber
Pillow==9.5.0
PyPDF2==3.0.1
spacy==3.8.3
python-pptx
pytesseract==0.3.13
pytest==8.3.5